Fix more cppcheck items

+++ rigs/tentec/tentec2.c
@@ -575,11 +575,11 @@ int tentec2_get_mode(RIG *rig, vfo_t vfo, rmode_t *mode, pbwidth_t *width)
 
     if (ttfilter < 16)
     {
-        *width = (ttfilter + 4) * 50;
+        *width = ((long)ttfilter + 4L) * 50L;
     }
     else
     {
-        *width = (ttfilter - 6) * 100;
+        *width = ((long)ttfilter - 6L) * 100L;
     }
 
     return RIG_OK;
